kernel-firmware: install brcmfmac 43430/43455 for all but RPi
brcmfmac43455-based hardware is now being used by x86_64 users, so we need to install the relevant firmwares from
linux-firmware, and limit deletion of these firmwares to only when building for RPi (which uses alternative and
possibly hardware specific firmware variants).

Hack in the missing 43430/43455 config files as Broadcom/Cypress don't see fit to publish these to linux-firmware. The
initial source for these files is the brcmfmac_sdio-firmware-rpi repo.

Ref: https://forum.libreelec.tv/thread/11249-how-to-change-files-on-read-only-file-system/?postID=104334#post104334